refactor types/index.ts
